Configure cog base URI via args to director
While we run director and model on the same network, it can be useful to
not have to. For example, when using Docker Compose to spin up a local
development environment, the model and director might use different
services. Being able to configure the endpoints will make setting that
up a lot simpler.